moved state.ts into models.ts
